Issue-21: Finished cleanup of main module.
