+++ app/build.gradle
@@ -29,6 +29,7 @@ android {
 dependencies {
     implementation libraries.appCompat
     implementation libraries.material
+    implementation libraries.coreKtx
 
     implementation project(':ranges')
 }
