--- app/src/main/java/com/example/kirchhoff/kotlin/delegation/User.kt
@@ -1,9 +1,0 @@
-package com.example.kirchhoff.kotlin.delegation
-
-/**
- * @author Kirchhoff-
- */
-class User(name: String, surname: String) {
-    val userName: String by NameDelegate(name)
-    val userSurname: String by NameDelegate(surname)
-}
