+++ app/src/main/java/com/example/kirchhoff/kotlin/example/intent/MainActivity.kt
@@ -1,0 +1,32 @@
+package com.example.kirchhoff.kotlin.example.intent
+
+import android.annotation.SuppressLint
+import android.app.Activity
+import android.content.Intent
+import android.os.Build
+import android.os.Bundle
+import androidx.core.view.isVisible
+import com.example.kirchhoff.kotlin.databinding.AOtherBinding
+import com.example.ranges.RangeActivity
+
+/**
+ * @author Kirchhoff-
+ */
+class MainActivity : Activity() {
+
+    private lateinit var binding: AOtherBinding
+
+    @SuppressLint("NewApi")
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = AOtherBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        with(binding) {
+            bRange.isVisible = Build.VERSION.SDK_INT >= 26
+            bRange.setOnClickListener {
+                startActivity(Intent(this@MainActivity, RangeActivity::class.java))
+            }
+        }
+    }
+}
