--- app/src/main/java/com/example/kirchhoff/kotlin/extensions/ViewExtensions.kt
@@ -1,7 +1,0 @@
-package com.example.kirchhoff.kotlin.extensions
-
-import android.view.View
-
-fun View.setVisibile(visible: Boolean) {
-    visibility = if (visible) View.VISIBLE else View.GONE
-}
