+++ app/src/main/res/layout/a_other.xml
@@ -5,13 +5,6 @@
     android:orientation="vertical">
 
     <Button
-        android:id="@+id/button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:text="GO!" />
-
-    <Button
         android:id="@+id/bRange"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
