+++ buildsystem/dependencies.gradle
@@ -9,7 +9,8 @@ ext.versions = [
         androidToolsBuildGradle: '7.2.1',
         kotlin                 : '1.6.21',
         appcompat              : '1.4.2',
-        material               : '1.6.1'
+        material               : '1.6.1',
+        coreKtx                : '1.8.0'
 ]
 
 ext.gradlePlugins = [
@@ -19,5 +20,6 @@ ext.gradlePlugins = [
 
 ext.libraries = [
         appCompat: "androidx.appcompat:appcompat:$versions.appcompat",
-        material : "com.google.android.material:material:$versions.material"
+        material : "com.google.android.material:material:$versions.material",
+        coreKtx: "androidx.core:core-ktx:$versions.coreKtx"
 ]
